Create HOG
This is what i have so far. I am still currently working on it because there is still some issue with it such as when the human player hit "y" to roll again, the turn goes to the computer instead of continuing the human's turn. Also it keeps prompting when human player hit "n"  
If you guys have any hints or edit it that would be great.
sorry for the delay on my part.